refactor: update log package and iam code to adjust requestID feature

+++ internal/apiserver/service/user_test.go
@@ -23,10 +23,10 @@ func TestMain(m *testing.M) {
 }
 
 func TestListUser(t *testing.T) {
-	var limit int = 3
+	var limit int64 = 3
 	opts := metav1.ListOptions{
-		Offset: pointer.ToInt(0),
-		Limit:  pointer.ToInt(limit),
+		Offset: pointer.ToInt64(0),
+		Limit:  pointer.ToInt64(limit),
 	}
 
 	got, err := ListUser(context.TODO(), opts)
@@ -38,7 +38,7 @@ func TestListUser(t *testing.T) {
 		t.Errorf("ListUser() TotalCount = %v, want %v", got.TotalCount, fake.ResourceCount)
 	}
 
-	if len(got.Items) != limit {
+	if len(got.Items) != int(limit) {
 		t.Errorf("len(UserListV2.Items)= %v, want %v", len(got.Items), limit)
 	}
 
@@ -57,8 +57,8 @@ func TestListUser(t *testing.T) {
 
 func BenchmarkListUser(b *testing.B) {
 	opts := metav1.ListOptions{
-		Offset: pointer.ToInt(0),
-		Limit:  pointer.ToInt(50),
+		Offset: pointer.ToInt64(0),
+		Limit:  pointer.ToInt64(50),
 	}
 
 	for i := 0; i < b.N; i++ {
